pyproject.toml
[tool.poetry]
package-mode = false
name = "todo-core"
version = "0.1.0"
description = ""
authors = ["Dmitriy <pilevichdima20@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pre-commit = "^4.0.1"
black = "^24.10.0"
drf-yasg = "^1.21.8"
celery = "^5.4.0"
python-decouple = "^3.8"
drf-nested-routers = "^0.94.1"
django-filter = "^24.3"
drf-spectacular = "^0.28.0"
redis = "^5.2.1"
aiokafka = "^0.12.0"
requests = "^2.32.3"
aiohttp = "^3.11.11"
flake8 = "^7.1.1"
asgiref = "^3.8.1"
dependency-injector = "^4.45.0"
httpx = "^0.28.1"
pillow = "^11.1.0"
confluent-kafka = "^2.8.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
